ci-requirements.txt
psycopg2<3.0
mysqlclient
Pillow
django-request-id
requests
flake8
codecov
coverage
django-hijack<3.0.0
openpyxl
pika
